// mamwebapp/src/app/models/asset-functional-performance.model.ts
export class AssetFunctionalPerformance {
    province: string;
    town: string;
    uniqueIdentifyingCode: string;
    possibleNonAssetSolutions: string;
    commonAssetDescription: string;
    currentUser: string;
    requiredPerformanceStandard: string;
    accessabilityRating: string;
    suitabilityIndex: string;
    conditionRating: string;
    operatingPerformanceIndex: string;
    functionalPerformanceRating: string;
}